# output/src/ApiApp/Controllers/OrderDetailsController.cs
// Generated by booster-apidev | 2026-04-06 19:37:30

using ApiApp.Data;
using ApiApp.DTOs.OrderDetails;
using ApiApp.Models;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;

namespace ApiApp.Controllers;

[ApiController]
[Route("api/orders/{orderId:int}/details")]
public class OrderDetailsController : ControllerBase
{
    private readonly AppDbContext _db;

    public OrderDetailsController(AppDbContext db)
    {
        _db = db;
    }

    // GET /api/orders/{orderId}/details
    [HttpGet]
    public async Task<ActionResult<IEnumerable<OrderDetailItemDto>>> GetOrderDetails(int orderId)
    {
        var orderExists = await _db.Orders.AnyAsync(o => o.Id == orderId);
        if (!orderExists) return NotFound(new { error = $"Order {orderId} not found." });

        var items = await _db.OrderDetails
            .Where(od => od.OrderId == orderId)
            .Include(od => od.Product)
            .Select(od => new OrderDetailItemDto
            {
                Id = od.Id,
                OrderId = od.OrderId,
                ProductId = od.ProductId,
                ProductCode = od.Product != null ? od.Product.ProductCode : null,
                ProductName = od.Product != null ? od.Product.ProductName : null,
                Quantity = od.Quantity,
                UnitPrice = od.UnitPrice,
                ExtendedPrice = od.Quantity * od.UnitPrice
            })
            .ToListAsync();

        return Ok(items);
    }

    // POST /api/orders/{orderId}/details
    [HttpPost]
    public async Task<ActionResult<OrderDetailItemDto>> AddOrderDetail(int orderId, [FromBody] OrderDetailCreateUpdateDto dto)
    {
        var orderExists = await _db.Orders.AnyAsync(o => o.Id == orderId);
        if (!orderExists) return NotFound(new { error = $"Order {orderId} not found." });

        var product = await _db.Products.FindAsync(dto.ProductId);
        if (product == null) return BadRequest(new { errors = new { ProductId = new[] { $"Product {dto.ProductId} not found." } } });

        decimal unitPrice = dto.UnitPrice ?? product.UnitPrice ?? 0m;

        var detail = new OrderDetail
        {
            OrderId = orderId,
            ProductId = dto.ProductId,
            Quantity = dto.Quantity,
            UnitPrice = unitPrice
        };

        _db.OrderDetails.Add(detail);
        await _db.SaveChangesAsync();

        return CreatedAtAction(nameof(GetOrderDetails), new { orderId }, new OrderDetailItemDto
        {
            Id = detail.Id,
            OrderId = detail.OrderId,
            ProductId = detail.ProductId,
            ProductCode = product.ProductCode,
            ProductName = product.ProductName,
            Quantity = detail.Quantity,
            UnitPrice = detail.UnitPrice,
            ExtendedPrice = detail.ExtendedPrice
        });
    }

    // PUT /api/orders/{orderId}/details/{id}
    [HttpPut("{id:int}")]
    public async Task<ActionResult<OrderDetailItemDto>> UpdateOrderDetail(int orderId, int id, [FromBody] OrderDetailCreateUpdateDto dto)
    {
        var detail = await _db.OrderDetails
            .Include(od => od.Product)
            .FirstOrDefaultAsync(od => od.Id == id && od.OrderId == orderId);

        if (detail == null) return NotFound(new { error = $"OrderDetail {id} not found in order {orderId}." });

        var product = await _db.Products.FindAsync(dto.ProductId);
        if (product == null) return BadRequest(new { errors = new { ProductId = new[] { $"Product {dto.ProductId} not found." } } });

        detail.ProductId = dto.ProductId;
        detail.Quantity = dto.Quantity;
        detail.UnitPrice = dto.UnitPrice ?? product.UnitPrice ?? 0m;
        detail.Product = product;

        await _db.SaveChangesAsync();

        return Ok(new OrderDetailItemDto
        {
            Id = detail.Id,
            OrderId = detail.OrderId,
            ProductId = detail.ProductId,
            ProductCode = product.ProductCode,
            ProductName = product.ProductName,
            Quantity = detail.Quantity,
            UnitPrice = detail.UnitPrice,
            ExtendedPrice = detail.ExtendedPrice
        });
    }

    // DELETE /api/orders/{orderId}/details/{id}
    [HttpDelete("{id:int}")]
    public async Task<IActionResult> DeleteOrderDetail(int orderId, int id)
    {
        var detail = await _db.OrderDetails
            .FirstOrDefaultAsync(od => od.Id == id && od.OrderId == orderId);

        if (detail == null) return NotFound(new { error = $"OrderDetail {id} not found in order {orderId}." });

        _db.OrderDetails.Remove(detail);
        await _db.SaveChangesAsync();
        return NoContent();
    }
}