# output/src/ApiApp/DTOs/Orders/OrderCreateDto.cs
// Generated by booster-apidev | 2026-04-06 19:37:30

namespace ApiApp.DTOs.Orders;

public class OrderCreateDto
{
    public int? CustomerId { get; set; }
    public int? EmployeeId { get; set; }
    public DateTime? OrderDate { get; set; }
    public DateTime? ShippedDate { get; set; }
    public DateTime? PaidDate { get; set; }
    public string? Status { get; set; }
    public string? Notes { get; set; }
}